app/members/serializers.py
from django.contrib.auth import get_user_model
from rest_framework import serializers

from members.models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = (
            'username',
            'email',
            'introduce',
        )