requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2018.10.15
chardet==3.0.4
decorator==4.3.0
Django==2.1.2
django-extensions==2.1.3
djangorestframework==3.9.0
idna==2.7
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.5
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
requests==2.20.0
Serializer==0.2.1
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
traitlets==4.3.2
urllib3==1.24
wcwidth==0.1.7
